+++ src/add-ons/kernel/bus_managers/pci/arch/x86/pci_bios.h
@@ -7,11 +7,11 @@ status_t pci_bios_init(void);
 
 status_t pci_bios_read_config(void *cookie, 
 							  uint8 bus, uint8 device, uint8 function,
-							  uint8 offset, uint8 size, uint32 *value);
+							  uint16 offset, uint8 size, uint32 *value);
 
 status_t pci_bios_write_config(void *cookie, 
 							   uint8 bus, uint8 device, uint8 function,
-							   uint8 offset, uint8 size, uint32 value);
+							   uint16 offset, uint8 size, uint32 value);
 
 status_t pci_bios_get_max_bus_devices(void *cookie, int32 *count);
